third_party/WebKit/LayoutTests/platform/win/virtual/service-worker-navigation-preload/http/tests/serviceworker/chromium/usecounter-on-load-expected.txt

This is a testharness.js-based test.
PASS usecounter is counted for a controlled document 
Harness: the test ran to completion.
